Moved TagsList and refactor for MUI
