+++ src/javascript/app/components/GalleryImage/index.scss
@@ -44,22 +44,6 @@
     align-content: flex-start;
   }
 
-  &__tag {
-    height: 18px;
-    border-radius: 9px;
-    font-size: 12px;
-    line-height: 14px;
-    margin: 0 3px 4px 3px;
-    padding: 2px 8px;
-    @include theme-background-color($color-secondary, $color-secondary);
-    @include theme-color($color-text-bright, $color-black);
-    overflow: hidden;
-
-    &--group {
-      @include theme-background-color($color-tertiary, $color-tertiary);
-    }
-  }
-
   &__created-meta {
     display: flex;
     justify-content: space-between;
